requirements.txt
asgiref==3.2.7
backcall==0.1.0
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
Django==3.0.5
django-cors-headers==3.2.1
django-rest-knox==4.1.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
jupyter-client==6.1.3
jupyter-core==4.6.3
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
pycparser==2.20
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pyzmq==19.0.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
